homework

+++ C# Basics/ConsoleInput-Output/01.SumOf3Numbers/SumOf3Numbers.cs
@@ -6,14 +6,14 @@ namespace _01.SumOf3Numbers
     {
         static void Main()
         {
-            // Write a program that reads 3 real numbers from the console and prints their sum.
+            // Write a program that reads 3 real numbers from the console and prdoubles their sum.
 
             Console.Write("a: ");
-            int a = int.Parse(Console.ReadLine());
+            double a = double.Parse(Console.ReadLine());
             Console.Write("b: ");
-            int b = int.Parse(Console.ReadLine());
+            double b = double.Parse(Console.ReadLine());
             Console.Write("c: ");
-            int c = int.Parse(Console.ReadLine());
+            double c = double.Parse(Console.ReadLine());
 
             Console.WriteLine("sum: " + (a + b + c));
