+++ C# Basics/ConsoleInput-Output/09.SumOfNNumbers/SumNNumbers.cs
@@ -6,15 +6,15 @@ namespace _09.SumOfNNumbers
     {
         static void Main()
         {
-            // Write a program that enters a number n and after that enters more n numbers and calculates and prints their sum. 
+            // Write a program that enters a number n and after that enters more n numbers and calculates and prdoubles their sum. 
             // Note that you may need to use a for-loop. 
 
             Console.Write("n: ");
-            int n = int.Parse(Console.ReadLine());
-            int sum = 0;
+            double n = double.Parse(Console.ReadLine());
+            double sum = 0.00;
             for (int i = 0; i < n; i++)
             {
-                int number = int.Parse(Console.ReadLine());
+                double number = double.Parse(Console.ReadLine());
                 sum += number;
             }
             Console.WriteLine("sum: " + sum);
